"#9 implemented roleEnum, rewrote tests"
