+++ OpenAPI/OpenAPI.json
@@ -122,7 +122,7 @@
             "description": "Role of account to get account role id",
             "required": true,
             "schema": {
-              "type": "string"
+              "$ref" : "contentSchema/enums/roleEnum.json"
             }
           }
         ],
