+++ OpenAPI/contentSchema/enums/roleEnum.json
@@ -1,0 +1,4 @@
+{
+    "type" : "string",
+    "enum" : ["client", "provider", "delivery", "manager"]
+}
